Separate renderer into their own classes
